# docs/swagger.yaml
definitions:
...
  models.URL:
    properties:
      contact_id:
        type: integer
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      type_label:
        type: string
      url:
        type: string
    type: object
  models.URLJSONPatch:
    properties:
      contact_id:
        example: 4
        type: integer
      id:
        example: 1
        type: integer
      is_primary:
        example: false
        type: boolean
      label_type_id:
        example: 42
        type: integer
      phone:
        example: https://facebook.com/kindredcard
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/steveredden/KindredCard/issues
  description: Personal CRM for managing contacts, relationships, and important dates
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.txt
  title: KindredCard API
  version: "1.0"
paths:
...
  /api/v1/contacts/{cid}/url:
    post:
      consumes:
      - application/json
      description: Associates a URL with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: integer
      - description: URL fields
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/models.URL'
      produces:
      - application/json
      responses:
        "200":
          description: 'created: newID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates a website / URL associated with a contact
      tags:
      - contacts
  /api/v1/contacts/{cid}/url/{uid}:
    delete:
      consumes:
      - application/json
      description: Associates a URL with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: URL ID
        in: path
        name: uid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates a website / URL associated with a contact
      tags:
      - contacts
...
  /api/v1/urls/{uid}:
    patch:
      consumes:
      - application/json
      description: Update specific fields of a url using HTTP PATCH. Only provided
        fields will be updated.
      parameters:
      - description: URL ID
        in: path
        name: uid
        required: true
        type: integer
      - description: URL fields to update
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.URLJSONPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Updated url
          schema:
            items:
              $ref: '#/definitions/models.URL'
            type: array
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Update a url number
      tags:
      - contacts
...
securityDefinitions:
  ApiTokenAuth:
    description: API token for authentication
    in: header
    name: session
    type: apiKey
swagger: "2.0"